+++ includes/feed-importing.php
@@ -227,13 +227,11 @@
 	function wprss_get_feed_items( $feed_url, $source, $force_feed = FALSE ) {
 		// Add filters and actions prior to fetching the feed items
 		add_filter( 'wp_feed_cache_transient_lifetime' , 'wprss_feed_cache_lifetime' );
-		add_action( 'wp_feed_options', 'wprss_do_not_cache_feeds' );
 
 		/* Fetch the feed from the soure URL specified */
 		$feed = wprss_fetch_feed( $feed_url, $source, $force_feed );
 
 		// Remove previously added filters and actions
-		remove_action( 'wp_feed_options', 'wprss_do_not_cache_feeds' );
 		remove_filter( 'wp_feed_cache_transient_lifetime' , 'wprss_feed_cache_lifetime' );
 
 		if ( !is_wp_error( $feed ) ) {
@@ -273,6 +271,47 @@ function wpse_cron_add_xdebug_cookie ($cron_request_array, $doing_wp_cron)
     return ($cron_request_array) ;
 }
 
+/**
+ * Checks if feed caching is enabled.
+ *
+ * This function will only return true if the cache option is enabled in the settings AND the cache directory exists
+ * or can be successfully created and is writable.
+ *
+ * @since [*next-version*]
+ *
+ * @return bool
+ */
+function wprss_is_feed_cache_enabled()
+{
+    $cacheOption = wprss_get_general_setting('feed_cache_enabled');
+    $cacheOption = filter_var($cacheOption, FILTER_VALIDATE_BOOLEAN);
+
+    if (!$cacheOption) {
+        return false;
+    }
+
+    $cacheDir = wprss_get_feed_cache_dir();
+
+    if (!file_exists($cacheDir)) {
+        @mkdir($cacheDir);
+    }
+
+    return is_writable($cacheDir);
+}
+
+/**
+ * Retrieves the feed cache directory.
+ *
+ * @since [*next-version*]
+ *
+ * @return string
+ */
+function wprss_get_feed_cache_dir()
+{
+    return wpra_container()->get('wpra/importer/cache/dir');
+}
+
+
 	/**
 	 * A clone of the function 'fetch_feed' in wp-includes/feed.php [line #529]
 	 *
@@ -308,7 +347,12 @@ function wpse_cron_add_xdebug_cookie ($cron_request_array, $doing_wp_cron)
         $fetch_time_limit = wprss_get_feed_fetch_time_limit();
         $feed->set_timeout($fetch_time_limit);
 
-        $feed->enable_cache(false);
+        $cacheEnabled = wprss_is_feed_cache_enabled();
+        $feed->enable_cache($cacheEnabled);
+
+        if ($cacheEnabled) {
+            $feed->set_cache_location(wprss_get_feed_cache_dir());
+        }
 
         // Reference array action hook, for the feed object and the URL
         do_action_ref_array('wp_feed_options', array(&$feed, $url));